+++ R/cms_product_details.r
@@ -16,24 +16,20 @@
 cms_product_details <- function(product, ...) {
   clients <- cms_get_client_info()
   stac_url <- gsub("/$", "", clients$catalogues[[1]]$stacRoot)
-  if (is.null(clients)) return(NULL) else {
-    product_url <- paste(
-      stac_url,
-      product,
-      "product.stac.json",
-      sep = "/"
-    )
-  }
-  
-  result <- .try_online({
+  product_url <- paste(
+    stac_url,
+    product,
+    "product.stac.json",
+    sep = "/"
+  )
+
+  result <-
     product_url |>
-      httr2::request() |>
-      httr2::req_perform()
-  }, "product-catalogue")
+    httr2::request() |>
+    httr2::req_perform() |>
+    httr2::resp_body_json()
   
-  if (is.null(result)) return(NULL) else {
-    result <- httr2::resp_body_json(result)
-    attr(result, "stac_url") <- stac_url
-    return(result)
-  }
+  attr(result, "stac_url") <- stac_url
+  
+  return(result)
 }