+++ R/cms_product_metadata.r
@@ -18,23 +18,15 @@
 #' @export
 cms_product_metadata <- function(product, ...) {
   details <- cms_product_details(product)
-  if (is.null(details)) return (NULL)
   links     <- lapply(details$links, as.data.frame) |> dplyr::bind_rows()
   item      <- links |> dplyr::filter(.data$rel == "item") |> dplyr::pull("href")
   meta_url  <- paste(attr(details, "stac_url"), product, item, sep = "/")
   result <-
     lapply(meta_url, function(u) {
-      .try_online({
-        u |>
-          httr2::request() |>
-          httr2::req_perform()
-      }, "meta-data-page")
+      httr2::request(u) |>
+        httr2::req_perform()
     }) |>
-    lapply(function(x) {
-      if (is.null(x)) return(NULL) else {
-        return(httr2::resp_body_json(x))
-      }
-    }) |>
+    lapply(httr2::resp_body_json) |>
     .simplify()
   return(result)
 }