+++ R/generics.r
@@ -1,40 +1,3 @@
-.http_status_ok <- function(x) {
-  if (dplyr::between(x$status_code, 100, 199)) {
-    message(sprintf("Unexpected informational response from Copernicus (status %i).", x$status_code))
-    return(FALSE)
-  }
-  if (dplyr::between(x$status_code, 300, 399)) {
-    message(sprintf("Unexpected redirection from Copernicus (status %i).", x$status_code))
-    return(FALSE)
-  }
-  if (dplyr::between(x$status_code, 400, 499)) {
-    message(sprintf(paste("Copernicus reported a client error (status %i).",
-                          "You may have requested information that is not available,",
-                          "please check your input.",
-                          sep = "\n"), x$status_code))
-    return(FALSE)
-  }
-  if (dplyr::between(x$status_code, 500, 599)) {
-    message(sprintf("Copernicus reported a server error (status %i).\nPlease try again later.", x$status_code))
-    return(FALSE)
-  }
-  if (x$status_code < 100 || x$status_code >= 600) {
-    message(sprintf("Copernicus responded with unknown status (status %i).", x$status_code))
-    return(FALSE)
-  }
-  return(TRUE)
-}
-
-.try_online <- function(expr, resource) {
-  result <- tryCatch(expr, error = function(e) {
-    message(sprintf("Failed to collect information from %s.\n%s", resource, e$message))
-    return(NULL)
-    })
-  if (is.null(result)) return(NULL)
-  if (!.http_status_ok(result)) return(NULL)
-  return(result)
-}
-
 .simplify <- function(data) {
   empty_row <- data.frame(a = NA)[,-1]
   result <-
@@ -61,4 +24,50 @@
         }
       )
     )
+}
+
+## Some requests result in a 403 status response, which trigger a warning.
+## Most likely this is the GDAL library trying to get directory listing
+## where the server does not allow it. These warnings are harmless and
+## do not affect the outcome. I will therefore muffle these warnings to
+## not confuse/bother the end-user.
+## Maybe this will be fixed in later GDAL releases.
+.muffle_403 <- function(expr) {
+  withCallingHandlers({
+    expr
+  }, warning = function(w) {
+    if (grepl(": 403", conditionMessage(w)))
+      invokeRestart("muffleWarning")
+  })
+}
+
+.uri_to_vsi <- function(href, progress, add_zarr = TRUE, streaming = TRUE) {
+  s3_root <- stringr::str_extract(href, "(?<=//)[^/]+")
+  check <-
+    Sys.setenv(AWS_S3_ENDPOINT     = s3_root) &&
+    Sys.setenv(AWS_NO_SIGN_REQUEST = "YES") &&
+    Sys.setenv(AWS_VIRTUAL_HOSTING = "FALSE")
+  
+  if (check) {
+    vsi <- href |>
+      stringr::str_replace("^https?://[^/]+/([^/]+)/(.*)$",
+                           sprintf("/vsis3%s/\\1/\\2",
+                                   ifelse(streaming, "_streaming", "")))
+  } else {
+    if (progress)
+      cli::cli_progress_message("Failed to set GDAL S3 config, trying alternative")
+    vsi <- paste0("/vsicurl/", href)
+  }
+  if (!add_zarr) return (vsi)
+  sprintf("ZARR::\"%s\"", vsi)
+}
+
+.get_stars_proxy <- function(vsi, variable) {
+  mdim_proxy <- .muffle_403({
+    stars::read_mdim(
+      vsi,
+      proxy = TRUE,
+      variable = variable
+    )
+  })
 }