+++ inst/doc/translate.R
@@ -26,8 +26,6 @@ translated <- cms_translate(cli_code)
 summary(translated)
 
 ## ----download-translation, message=FALSE, fig.width=4, fig.height=4, fig.alt="Data downloaded using translated query code"----
-if (requireNamespace("blosc")) {
-  result <- do.call(cms_download_subset, translated)
-  plot(result, col = hcl.colors(100), axes = TRUE)
-}
+result <- do.call(cms_download_subset, translated)
+plot(result, col = hcl.colors(100), axes = TRUE)