+++ tests/testthat/test_login.r
@@ -10,6 +10,11 @@ test_that("Login works", {
 test_that("Setting username works", {
   skip_on_cran()
   expect_true({
+    curuid <- cms_get_username()
+    on.exit({
+      cms_set_username(curuid, "sysenv")
+      cms_set_username(curuid, "option")
+    }, add = TRUE, after = FALSE)
     cms_set_username("henk", "option")
     cms_set_username("henk", "sysenv")
     cms_get_username() == "henk"
@@ -19,19 +24,32 @@ test_that("Setting username works", {
 test_that("Setting password works", {
   skip_on_cran()
   expect_true({
+    curpwd <- cms_get_password()
+    on.exit({
+      cms_set_password(curpwd, "sysenv")
+      cms_set_password(curpwd, "option")
+    }, add = TRUE, after = FALSE)
     cms_set_password("foobar", "option")
     cms_set_password("foobar", "sysenv")
     cms_get_password() == "foobar"
   })
 })
 
-test_that("Using dummy account details returns NULL", {
+test_that("Using dummy account details throws error", {
   skip_on_cran()
-  expect_null({
+  expect_error({
+    curuid <- cms_get_username()
+    curpwd <- cms_get_password()
+    on.exit({
+      cms_set_username(curuid, "sysenv")
+      cms_set_username(curuid, "option")
+      cms_set_password(curpwd, "sysenv")
+      cms_set_password(curpwd, "option")
+    }, add = TRUE, after = FALSE)
     cms_set_username("henk", "option")
     cms_set_username("henk", "sysenv")
     cms_set_password("foobar", "option")
     cms_set_password("foobar", "sysenv")
     cms_login() |> suppressMessages()
-  })
+  }, "401 Unauthorized")
 })